implemented leo's comments
