+++ src/pil/backend/mod.rs
@@ -1,0 +1,1 @@
+pub mod powdr_pil;
