+++ src/pil/compiler/mod.rs
@@ -1,0 +1,1 @@
+pub mod powdr_pil;
