+++ src/plonkish/backend/mod.rs
@@ -1,3 +1,2 @@
 pub mod halo2;
 pub mod plaf;
-pub mod powdr_pil;
